build_all for nvpro_core2
Not all samples have been ported to nvpro_core2 yet -- more will
over time and we're hoping to port or archive all of them -- so in
this interim period, build_all lets you build both v1 and v2 samples
in the same project (default behavior), just v1 samples (run CMake
in the legacy/ folder), or just v2 samples (add
`-DNVPRO_SAMPLES_NO_V1_SAMPLES=ON` to your command line).

We've also moved the lists of all samples that appeared in the
various batch scripts to a single .csv file. This file is shared
between both the scripts and CMake. v1 samples are listed with a
",v1" at the end of their name; anything else is a v2 sample.

Once all the samples use nvpro_core2 (or are archived with the
last copies of nvpro_core 1 they work with added as submodules to
their projects), we'll remove the `legacy` folder, and the .csv
file will just be a list of project names.